pyproject.toml
[project]
name = "app-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Robert Hofer", email = "1058012+hofrob@users.noreply.github.com" }
]
requires-python = ">=3.14"
dependencies = [
    "alembic[tz]>=1.14.0",
    "asyncpg>=0.30.0",
    "faker>=33.0.0",
    "fastapi-slim>=0.115.5",
    "psycopg[binary]",
    "sqlalchemy-utils>=0.41.2",
    "sqlalchemy[asyncio]>=2.0.36",
    "uvicorn[standard]>=0.32.1",
]
[dependency-groups]
dev = [
    "black",
    "mypy",
    "pre-commit",
    "ruff",
]

[project.scripts]
app-poc = "app_poc:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
target-version = [
    'py314',
]
skip-magic-trailing-comma = true

[tool.ruff]
target-version = "py314"
line-length = 88
extend-exclude = [
    "temp",
]
preview = true
fix = true
lint.select = [
    "ALL",
]
lint.ignore = [
    "A005",    # Module is shadowing a Python builtin module
    "C901",    # function-is-too-complex
    "COM812",  # Trailing comma missing: incompatible with black
    "CPY001",  # missing-copyright-notice
    "D100",    # Missing docstring in public module
    "D101",    # Missing docstring in public class
    "D102",    # Missing docstring in public method
    "D103",    # Missing docstring in public function
    "D104",    # Missing docstring in public package
    "D105",    # Missing docstring in magic method
    "D106",    # Missing docstring in public nested class
    "D107",    # Missing docstring in __init__
    "D203",    # 1 blank line required before class docstring
    "D205",    # 1 blank line required between summary line and description
    "D213",    # Multi-line docstring summary should start at the second line
    "DOC201",  # return is not documented in docstring
    "DOC501",  # Raised exception missing from docstring
    "PLR0904", # Too many public methods
    "PLR0913", # Too many arguments in function definition
    "PLR2004", # Magic value used in comparison, consider replacing {value} with a constant variable
    "PLW3201", # Bad or misspelled dunder method name
    "TC001",
    "TC002",
    "TC003",
]
lint.per-file-ignores."__init__.py" = [
    "F401",
]
lint.flake8-import-conventions.aliases.sqlalchemy = "sa"
lint.isort.combine-as-imports = true
lint.isort.force-sort-within-sections = true
lint.isort.split-on-trailing-comma = false
lint.pep8-naming.classmethod-decorators = [
    "classmethod",
    "pydantic.field_validator",
]
